requirements.txt
streamlit>=1.50.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
yfinance>=0.2.18
pyyaml>=6.0
requests>=2.28.0
scikit-learn>=1.3.0
ta-lib>=0.4.25